view: added the designed for the cards

+++ lib/config/database/entities/entity_cashflow.dart
@@ -2,8 +2,8 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 
 class EntityCashflow {
   String? _id;
-  int? _openValue;
-  int? _closeValue;
+  double? _openValue;
+  double? _closeValue;
   bool? _isOpen;
   Timestamp? _init;
   Timestamp? _end;
@@ -11,7 +11,7 @@ class EntityCashflow {
   EntityCashflow(
       {required String id,
       required Timestamp init,
-      required int openValue,
+      required double openValue,
       required bool isOpen})
       : _id = id,
         _openValue = openValue,
@@ -19,8 +19,8 @@ class EntityCashflow {
         _isOpen = isOpen;
 
   String getId() => _id!;
-  int getOpenValue() => _openValue!;
-  int getCloseValue() => _closeValue!;
+  double getOpenValue() => _openValue!;
+  double getCloseValue() => _closeValue!;
   bool getState() => _isOpen!;
   Timestamp getInitTime() => _init!;
   String getInitTimeToString() {
@@ -29,8 +29,8 @@ class EntityCashflow {
   }
 
   Timestamp getEndTime() => _end!;
-  setOpenValue(int value) => _openValue = value;
-  setCloseValue(int value) => _closeValue = value;
+  setOpenValue(double value) => _openValue = value;
+  setCloseValue(double value) => _closeValue = value;
   setState(bool value) => _isOpen = value;
   setInitTime(Timestamp value) => _init = value;
   setCloseTime(Timestamp value) => _end = value;
